Python: move chop + resource into rst_code_example_pipeline (git mv)
Move sphinx/widget/chop.py and sphinx/widget/resource.py into the
package via git mv (history preserved):
  frontend/python/rst_code_example_pipeline/src/rst_code_example_pipeline/

chop.py's `from .resource import Resource` is already a relative import
and continues to work in its new location within the package.

Update extract_projects.py: `from widget.chop import …` → `from .chop import …`
— the cross-package import is now an intra-package relative import.

Drop `extraPaths = ["../../sphinx"]` from [tool.pyright] in pyproject.toml;
the package no longer has any external dependency on sphinx/.
