+++ frontend/python/rst_code_example_pipeline/pyproject.toml
@@ -17,4 +17,3 @@ where = ["src"]
 
 [tool.pyright]
 pythonVersion = "3.10"
-extraPaths = ["../../sphinx"]
