+++ frontend/python/rst_code_example_pipeline/src/rst_code_example_pipeline/chop.py
